+++ modules/markup/sanitizer.go
@@ -3,7 +3,7 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-package markdown
+package markup
 
 import (
 	"regexp"
